+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ## Unreleased
+### Changed
+- BREAKING: Changed batch.readFile() to pass parsed announcements instead of parquet records to callback
+- BREAKING: Changed createdAt on announcements to BigInt instead of number
 
 ## [2.1.1] - 2021-08-17
 ### Changed